dvfs: Move dentry structure to separate file
